requirements.txt
# -----------------------------------------------------------------------------
# requirements.txt - المكتبات المطلوبة للمشروع
# -----------------------------------------------------------------------------

python-telegram-bot==21.0.1
# سنضيف مكتبات بينانس والتحليل لاحقًا
# python-binance
# pandas-ta

